+++ CouponSystemSpring/src/main/java/com/mbms/epository/CouponRepository.java
@@ -9,7 +9,7 @@ import org.springframework.data.jpa.repository.Query;
 import com.mbms.model.Coupon;
 import com.mbms.model.CouponCaregory;
 
-public interface CouponRepository extends JpaRepository<Coupon, Integer>{
+public interface CouponRepository extends JpaRepository<Coupon, Long>{
 	
 	Coupon findByTitle(String title);
 
@@ -19,18 +19,13 @@ public interface CouponRepository extends JpaRepository<Coupon, Integer>{
 
 		List<Coupon> findByPriceLessThan(double price);
 
-	//	List<Coupon> findByEndDateBefore(Date date);
-
 		@Query("select c from Coupon c where c.id = ?1  AND  c.company.id = ?2")
-
 		Coupon getCouponCompany(int couponId, int companyId);
 
 		@Query("SELECT DISTINCT c FROM Coupon c INNER JOIN c.customers t where t.id = ?1")
-
 		List<Coupon> couponsCustomerByCustomerId(int customerId);
 
 		@Query("SELECT DISTINCT c FROM Coupon c INNER JOIN c.customers t where t.id = ?1 and c.id = ?2")
-
 		Coupon couponByCustomerIdAndCouponId(int customerId, int couponId);
 
 		List<Coupon> findByEndDateBefore(java.util.Date date);
